samantha-ai-assistant/backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.8.0
eventlet==0.33.3
psycopg2-binary==2.9.7
redis==4.6.0
openai==0.28.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
whisper==1.1.10
numpy==1.24.3
torch==2.0.1
torchaudio==2.0.2
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
